+++ stdnet/__init__.py
@@ -24,7 +24,7 @@ CLASSIFIERS = [
                ]
 
 
-sphinxtogithub = False
+sphinxtogithub = True
 
 from .exceptions import *
 from .backends import *