+++ CHANGELOG.rst
@@ -1,5 +1,6 @@
 Ver. 0.5.5 - Development
 ============================
+* Added :func:`stdnet.orm.flush_models`, a utility functions for flushing model data.
 * Added a new :class:`stdnet.orm.ByteField` which saves bytes rather than strings.
 * Renamed ``start`` and ``end`` in TimeSeres and ``data_start`` and ``data_end``.